feat: 공모 내역 테이블 컴포넌트 별도 분리 및 데이터 연결
